# contrato/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "description" : "Inventario, ventas, cotizaciones e instalaciones de ITALARM. Los errores responden en Problem Details (RFC 9457) con la propiedad `codigo`. El dinero viaja como texto decimal junto con su moneda.",
    "title" : "API ITALARM",
    "version" : "v1"
  },
  "servers" : [ {
    "url" : "/"
  } ],
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
...
  "paths" : {
...
    "/api/v1/clientes/{id}" : {
      "get" : {
        "operationId" : "detalle_2",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ClienteVista"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Detalle de un cliente",
        "tags" : [ "Clientes" ]
      },
      "put" : {
        "description" : "Además: MODIFICADO_POR_OTRO_USUARIO (409).",
        "operationId" : "actualizar_4",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SolicitudCliente"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ClienteVista"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Editar un cliente",
        "tags" : [ "Clientes" ]
      }
    },
    "/api/v1/compras" : {
      "get" : {
        "description" : "Sin fechas, el mes en curso. Los totales no incluyen las anuladas. Orden por defecto: más recientes primero.",
        "operationId" : "listar_2",
        "parameters" : [ {
          "in" : "query",
          "name" : "proveedorId",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "in" : "query",
          "name" : "productoId",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "in" : "query",
          "name" : "desde",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date"
          }
        }, {
          "in" : "query",
          "name" : "hasta",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date"
          }
        }, {
          "description" : "Incluir las anuladas en el listado (por defecto, sí)",
          "in" : "query",
          "name" : "incluirAnuladas",
          "required" : false,
          "schema" : {
            "type" : "boolean"
          }
        }, {
          "description" : "Zero-based page index (0..N)",
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "default" : 0,
            "minimum" : 0
          }
        }, {
          "description" : "The size of the page to be returned",
          "in" : "query",
          "name" : "size",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "default" : 20,
            "minimum" : 1
          }
        }, {
          "description" : "Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
          "in" : "query",
          "name" : "sort",
          "required" : false,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ListadoComprasVista"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Listado paginado de compras con los totales del período",
        "tags" : [ "Compras" ]
      },
      "post" : {
        "description" : "Con la misma Idempotency-Key devuelve la compra ya creada. Errores (400): SERIALES_NO_COINCIDEN, CANTIDAD_INVALIDA, COMPRA_PRODUCTO_REPETIDO, COMPRA_FECHA_FUTURA, COMPRA_COSTO_INVALIDO, PROVEEDOR_NO_EXISTE, PRODUCTO_NO_EXISTE; (409) SERIAL_DUPLICADO; (422) TASA_NO_DISPONIBLE, PRODUCTO_INACTIVO.",
        "operationId" : "registrar",
        "parameters" : [ {
          "description" : "Clave única por intento de guardar (RT-07)",
          "in" : "header",
          "name" : "Idempotency-Key",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SolicitudCompra"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CompraVista"
                }
              }
            },
            "description" : "Created"
          }
        },
        "summary" : "Registrar una compra y sumarla al inventario",
        "tags" : [ "Compras" ]
      }
    },
    "/api/v1/compras/vista-previa" : {
      "post" : {
        "description" : "Es el valor oficial que muestra el frontend (BF-06). Errores: TASA_NO_DISPONIBLE, COMPRA_PRODUCTO_REPETIDO, COMPRA_FECHA_FUTURA, PRODUCTO_NO_EXISTE.",
        "operationId" : "vistaPrevia_1",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SolicitudVistaPreviaCompra"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/VistaPreviaCompraVista"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Vista previa: costo actual → nuevo y regla por línea, sin guardar",
        "tags" : [ "Compras" ]
      }
    },
    "/api/v1/compras/{id}" : {
      "get" : {
        "operationId" : "detalle_4",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CompraVista"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Detalle de una compra, con si se puede anular y por qué no",
        "tags" : [ "Compras" ]
      }
    },
    "/api/v1/compras/{id}/anular" : {
      "post" : {
        "description" : "Solo si es el último movimiento de cada producto y sus seriales siguen en bodega. Errores: COMPRA_NO_ANULABLE (422), COMPRA_YA_ANULADA (409).",
        "operationId" : "anular",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SolicitudAnulacion"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CompraVista"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Anular una compra",
        "tags" : [ "Compras" ]
      }
    },
    "/api/v1/compras/{id}/factura" : {
      "delete" : {
        "operationId" : "quitarFactura",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CompraVista"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Quitar la factura",
        "tags" : [ "Compras" ]
      },
      "put" : {
        "description" : "Errores: ARCHIVO_TIPO_NO_PERMITIDO, ARCHIVO_DEMASIADO_GRANDE (400).",
        "operationId" : "cambiarFactura",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "multipart/form-data" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "archivo" : {
                    "type" : "string",
                    "format" : "binary"
                  }
                },
                "required" : [ "archivo" ]
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CompraVista"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Subir o reemplazar la factura (JPEG, PNG, WebP o PDF; máximo 5 MB)",
        "tags" : [ "Compras" ]
      }
    },
...
    "/api/v1/productos/{id}" : {
      "delete" : {
        "description" : "Error: PRODUCTO_CON_MOVIMIENTOS (422).",
        "operationId" : "eliminar_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        },
        "summary" : "Eliminar un producto sin movimientos",
        "tags" : [ "Productos" ]
      },
      "get" : {
        "operationId" : "detalle_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProductoVista"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Detalle de un producto",
        "tags" : [ "Productos" ]
      },
      "put" : {
        "description" : "Errores: PRODUCTO_CODIGO_DUPLICADO, MODIFICADO_POR_OTRO_USUARIO (409); PRODUCTO_CAMBIO_NO_PERMITIDO (422).",
        "operationId" : "actualizar_2",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SolicitudProducto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProductoVista"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Editar un producto",
        "tags" : [ "Productos" ]
      }
    },
...
    "/api/v1/proveedores/{id}" : {
      "get" : {
        "operationId" : "detalle",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProveedorVista"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Detalle de un proveedor",
        "tags" : [ "Proveedores" ]
      },
      "put" : {
        "description" : "Además: MODIFICADO_POR_OTRO_USUARIO (409).",
        "operationId" : "actualizar_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SolicitudProveedor"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProveedorVista"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Editar un proveedor",
        "tags" : [ "Proveedores" ]
      }
    },
...
    "/api/v1/tasas/vista-previa" : {
      "post" : {
        "operationId" : "vistaPrevia",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SolicitudVistaPrevia"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/VariacionVista"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Tasa anterior, nueva y variación antes de guardar (RF-35b); no guarda nada",
        "tags" : [ "Tasas de cambio" ]
      }
    },
    "/api/v1/tasas/{id}" : {
      "get" : {
        "operationId" : "detalle_3",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TasaVista"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Detalle de una tasa con sus correcciones",
        "tags" : [ "Tasas de cambio" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "ClienteVista" : {
        "type" : "object",
        "properties" : {
          "cantidadMovimientos" : {
            "type" : "integer",
            "format" : "int64"
          },
          "ciudad" : {
            "type" : "string"
          },
          "correo" : {
            "type" : "string"
          },
          "direccion" : {
            "type" : "string"
          },
          "fechaUltimoMovimiento" : {
            "type" : "string",
            "format" : "date"
          },
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "nombre" : {
            "type" : "string"
          },
          "numeroDocumento" : {
            "type" : "string"
          },
          "precioAplicado" : {
            "type" : "string",
            "enum" : [ "INSTALADOR", "CLIENTE_FINAL" ]
          },
          "precioAplicadoDescripcion" : {
            "type" : "string"
          },
          "telefono" : {
            "type" : "string"
          },
          "tipo" : {
            "type" : "string",
            "enum" : [ "INSTALADOR", "CLIENTE_FINAL" ]
          },
          "tipoDocumento" : {
            "type" : "string",
            "enum" : [ "CC", "NIT" ]
          },
          "version" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
...
      "CompraVista" : {
        "type" : "object",
        "properties" : {
          "anulable" : {
            "type" : "boolean"
          },
          "anulacion" : {
            "$ref" : "#/components/schemas/Anulacion"
          },
          "consecutivo" : {
            "type" : "string"
          },
          "estado" : {
            "type" : "string"
          },
          "facturaUrl" : {
            "type" : "string"
          },
          "fecha" : {
            "type" : "string",
            "format" : "date"
          },
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "lineas" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Linea"
            }
          },
          "moneda" : {
            "type" : "string",
            "enum" : [ "USD", "COP", "VES" ]
          },
          "motivoNoAnulable" : {
            "type" : "string"
          },
          "numeroFactura" : {
            "type" : "string"
          },
          "proveedor" : {
            "$ref" : "#/components/schemas/Referencia"
          },
          "registradaEn" : {
            "type" : "string",
            "format" : "date-time"
          },
          "registradaPor" : {
            "type" : "string"
          },
          "tasas" : {
            "$ref" : "#/components/schemas/TasasCompraVista"
          },
          "total" : {
            "$ref" : "#/components/schemas/Dinero"
          },
          "totalUsd" : {
            "$ref" : "#/components/schemas/Dinero"
          }
        }
      },
...
      "Dinero" : {
        "type" : "object",
        "properties" : {
          "moneda" : {
            "type" : "string",
            "enum" : [ "USD", "COP", "VES" ]
          },
          "monto" : {
            "type" : "string",
            "format" : "decimal",
            "example" : "19.5000",
            "pattern" : "^-?\\d+(\\.\\d+)?$"
          }
        }
      },
...
      "ListadoComprasVista" : {
        "type" : "object",
        "properties" : {
          "compras" : {
            "$ref" : "#/components/schemas/PaginaCompraResumenVista"
          },
          "desde" : {
            "type" : "string",
            "format" : "date"
          },
          "hasta" : {
            "type" : "string",
            "format" : "date"
          },
          "totalUsd" : {
            "$ref" : "#/components/schemas/Dinero"
          },
          "totalesPorMoneda" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/TotalMoneda"
            }
          }
        }
      },
...
      "ProductoVista" : {
        "type" : "object",
        "properties" : {
          "activo" : {
            "type" : "boolean"
          },
          "bajoMinimo" : {
            "type" : "boolean"
          },
          "categoria" : {
            "$ref" : "#/components/schemas/Referencia"
          },
          "codigo" : {
            "type" : "string"
          },
          "controlaSerial" : {
            "type" : "boolean"
          },
          "costoActual" : {
            "$ref" : "#/components/schemas/Dinero"
          },
          "descripcion" : {
            "type" : "string"
          },
          "fotoUrl" : {
            "type" : "string"
          },
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "marca" : {
            "type" : "string"
          },
          "modelo" : {
            "type" : "string"
          },
          "nombre" : {
            "type" : "string"
          },
          "precioClienteFinal" : {
            "$ref" : "#/components/schemas/Dinero"
          },
          "precioInstalador" : {
            "$ref" : "#/components/schemas/Dinero"
          },
          "stock" : {
            "type" : "string",
            "format" : "decimal",
            "example" : "19.5000",
            "pattern" : "^-?\\d+(\\.\\d+)?$"
          },
          "stockMinimo" : {
            "type" : "string",
            "format" : "decimal",
            "example" : "19.5000",
            "pattern" : "^-?\\d+(\\.\\d+)?$"
          },
          "unidadMedida" : {
            "$ref" : "#/components/schemas/UnidadMedidaVista"
          },
          "version" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
      "ProveedorVista" : {
        "type" : "object",
        "properties" : {
          "ciudad" : {
            "type" : "string"
          },
          "correo" : {
            "type" : "string"
          },
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "monedaHabitual" : {
            "type" : "string",
            "enum" : [ "USD", "COP", "VES" ]
          },
          "nit" : {
            "type" : "string"
          },
          "nombre" : {
            "type" : "string"
          },
          "telefono" : {
            "type" : "string"
          },
          "version" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
...
      "SolicitudAnulacion" : {
        "type" : "object",
        "properties" : {
          "motivo" : {
            "type" : "string",
            "maxLength" : 300,
            "minLength" : 0
          }
        },
        "required" : [ "motivo" ]
      },
...
      "SolicitudCliente" : {
        "type" : "object",
        "properties" : {
          "ciudad" : {
            "type" : "string",
            "maxLength" : 80,
            "minLength" : 0
          },
          "correo" : {
            "type" : "string",
            "format" : "email",
            "maxLength" : 254,
            "minLength" : 0
          },
          "direccion" : {
            "type" : "string",
            "maxLength" : 200,
            "minLength" : 0
          },
          "nombre" : {
            "type" : "string",
            "maxLength" : 150,
            "minLength" : 0
          },
          "numeroDocumento" : {
            "type" : "string",
            "maxLength" : 30,
            "minLength" : 0,
            "pattern" : "[0-9.\\s-]*"
          },
          "telefono" : {
            "type" : "string",
            "maxLength" : 25,
            "minLength" : 0
          },
          "tipo" : {
            "type" : "string",
            "enum" : [ "INSTALADOR", "CLIENTE_FINAL" ]
          },
          "tipoDocumento" : {
            "type" : "string",
            "enum" : [ "CC", "NIT" ]
          },
          "version" : {
            "type" : "integer",
            "format" : "int64"
          }
        },
        "required" : [ "nombre", "telefono", "tipo", "version" ]
      },
      "SolicitudCompra" : {
        "type" : "object",
        "properties" : {
          "fecha" : {
            "type" : "string",
            "format" : "date"
          },
          "lineas" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/SolicitudLineaCompra"
            },
            "maxItems" : 200,
            "minItems" : 0
          },
          "moneda" : {
            "type" : "string",
            "enum" : [ "USD", "COP", "VES" ]
          },
          "numeroFactura" : {
            "type" : "string",
            "maxLength" : 50,
            "minLength" : 0
          },
          "proveedorId" : {
            "type" : "integer",
            "format" : "int64"
          }
        },
        "required" : [ "lineas", "moneda", "numeroFactura", "proveedorId" ]
      },
...
      "SolicitudProducto" : {
        "type" : "object",
        "properties" : {
          "categoriaId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "codigo" : {
            "type" : "string",
            "maxLength" : 30,
            "minLength" : 0,
            "pattern" : "\\s*[A-Za-z0-9][A-Za-z0-9._/-]*\\s*"
          },
          "controlaSerial" : {
            "type" : "boolean"
          },
          "descripcion" : {
            "type" : "string",
            "maxLength" : 2000,
            "minLength" : 0
          },
          "marca" : {
            "type" : "string",
            "maxLength" : 80,
            "minLength" : 0
          },
          "modelo" : {
            "type" : "string",
            "maxLength" : 80,
            "minLength" : 0
          },
          "monedaPrecio" : {
            "type" : "string",
            "enum" : [ "USD", "COP", "VES" ]
          },
          "nombre" : {
            "type" : "string",
            "maxLength" : 150,
            "minLength" : 0
          },
          "precioClienteFinal" : {
            "type" : "string",
            "format" : "decimal",
            "example" : "19.5000",
            "pattern" : "^-?\\d+(\\.\\d+)?$"
          },
          "precioInstalador" : {
            "type" : "string",
            "format" : "decimal",
            "example" : "19.5000",
            "pattern" : "^-?\\d+(\\.\\d+)?$"
          },
          "stockMinimo" : {
            "type" : "string",
            "format" : "decimal",
            "example" : "19.5000",
            "pattern" : "^-?\\d+(\\.\\d+)?$"
          },
          "unidadMedidaId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "version" : {
            "type" : "integer",
            "format" : "int64"
          }
        },
        "required" : [ "categoriaId", "codigo", "controlaSerial", "nombre", "precioClienteFinal", "precioInstalador", "unidadMedidaId", "version" ]
      },
...
      "SolicitudVistaPrevia" : {
        "type" : "object",
        "properties" : {
          "par" : {
            "type" : "string",
            "enum" : [ "USD_COP", "USD_VES" ]
          },
          "tasaId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "valor" : {
            "type" : "string",
            "format" : "decimal",
            "example" : "19.5000",
            "pattern" : "^-?\\d+(\\.\\d+)?$"
          }
        },
        "required" : [ "par", "valor" ]
      },
      "SolicitudVistaPreviaCompra" : {
        "type" : "object",
        "properties" : {
          "fecha" : {
            "type" : "string",
            "format" : "date"
          },
          "lineas" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/SolicitudLineaCompra"
            },
            "maxItems" : 200,
            "minItems" : 0
          },
          "moneda" : {
            "type" : "string",
            "enum" : [ "USD", "COP", "VES" ]
          }
        },
        "required" : [ "lineas", "moneda" ]
      },
...
      "VistaPreviaCompraVista" : {
        "type" : "object",
        "properties" : {
          "avisos" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "fecha" : {
            "type" : "string",
            "format" : "date"
          },
          "lineas" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Linea"
            }
          },
          "moneda" : {
            "type" : "string",
            "enum" : [ "USD", "COP", "VES" ]
          },
          "tasas" : {
            "$ref" : "#/components/schemas/TasasCompraVista"
          },
          "total" : {
            "$ref" : "#/components/schemas/MontoEnMonedas"
          }
        }
      }
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "description" : "Token recibido en POST /api/v1/sesion.",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  }
}